oops: stub out with basic implementation to support gopherjs
