+++ oops/xerrors.go
@@ -1,3 +1,4 @@
+// +build !js
 // Copyright (c) 2019 The Go Authors. All rights reserved.
 
 // Redistribution and use in source and binary forms, with or without
